Restructured the folders and files
