+++ src/components/Pages/Tasks/TaskRow/EditableName.tsx
@@ -1,6 +1,6 @@
 import { useEffect, useRef, useState } from "react"
-import usePopover from "../../hooks/usePopover";
-import Popover from "../Shared Components/Popover";
+import usePopover from "../../../../hooks/usePopover";
+import Popover from "../../../Shared Components/Popover";
 
 /**
  * This variable will set the limit for the text input in the component. After the limit is reached it will display a warning to the user.
